\name{thermometers2004}
\alias{thermometers2004}
\docType{data}
\title{Thermometer Ratings from the 2004 CPS American National Election Study}
\description{These data are thermometer ratings from the 2004 CPS American National Election Study.  It is a numeric matrix that has, in each column, individual thermometer ratings for 13 political figures and parties prior to the 2004 election.}
\usage{data(thermometers2004)}
\format{The 13 columns correspond to ratings for: George W. Bush, John Kerry, Ralph Nader, Dick Cheney, John Edwards, Laura Bush, Hilary Clinton, Bill Clinton, Colin Powell, John Ashcroft, John McCain, the Democratic Party and the Republican Party, respectively.}
\examples{
data(thermometers2004)
}
\keyword{datasets}
